FastNoise2 & Other Stuff Added

+++ modules/math/math.cpp
@@ -7,15 +7,15 @@
 
 void InitializeMathModule()
 {
-	Gin::Module::AddNodesToRegistry<Gin::Module::Math::Multiply<double>>("Math/Multiply");
-	Gin::Module::AddNodesToRegistry<Gin::Module::Math::Add<double>>("Math/Add");
-	Gin::Module::AddNodesToRegistry<Gin::Module::Math::Substract<double>>("Math/Substract");
-	Gin::Module::AddNodesToRegistry<Gin::Module::Math::Divide<double>>("Math/Divide");
+	Gin::Module::AddNodesToRegistry<Gin::Module::Math::Multiply<float>>("Math/Multiply");
+	Gin::Module::AddNodesToRegistry<Gin::Module::Math::Add<float>>("Math/Add");
+	Gin::Module::AddNodesToRegistry<Gin::Module::Math::Substract<float>>("Math/Substract");
+	Gin::Module::AddNodesToRegistry<Gin::Module::Math::Divide<float>>("Math/Divide");
 
-	Gin::Module::AddNodesToRegistry<Gin::Module::Math::MultiplySpatial<double>>("Math/MultiplySpatial");
-	Gin::Module::AddNodesToRegistry<Gin::Module::Math::AddSpatial<double>>("Math/AddSpatial");
-	Gin::Module::AddNodesToRegistry<Gin::Module::Math::SubstractSpatial<double>>("Math/SubstractSpatial");
-	Gin::Module::AddNodesToRegistry<Gin::Module::Math::DivideSpatial<double>>("Math/DivideSpatial");
+	Gin::Module::AddNodesToRegistry<Gin::Module::Math::MultiplySpatial<float>>("Math/MultiplySpatial");
+	Gin::Module::AddNodesToRegistry<Gin::Module::Math::AddSpatial<float>>("Math/AddSpatial");
+	Gin::Module::AddNodesToRegistry<Gin::Module::Math::SubstractSpatial<float>>("Math/SubstractSpatial");
+	Gin::Module::AddNodesToRegistry<Gin::Module::Math::DivideSpatial<float>>("Math/DivideSpatial");
 }
 
 void UninitializeMathModule()
